Move dtw logic to dtw module
